bert_vis/requirements.txt
bokeh
bert-serving-client
umap-learn